share/install: fix broken installs from --reconfigure
There was an old-school thought from a previous generation of the
program that --reconfigure should enforce a configured state. The
install process has since become a little more complex, and
--reconfigure mode never had the additional steps added.

This behavior isn't very desirable anyhow. The program will behave more
robustly if --reconfigure never attempts to run for an uninstalled
package.